+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Reset `.has_extra_trafo` to `FALSE` when trafo is set to `NULL`.
 * `rd_info.ParamSet` collapses vector with `"\n"` due changes in roxygen 7.2.0
+* Add method `set_values()` to conveniently add parameter values.
 
 # paradox 0.9.0